refatorando

+++ src/FleetManager.Domain/Entities/Vehicle.cs
@@ -14,11 +14,11 @@ namespace FleetManager.Domain.Entities
         public string LicensePlate { get; set; } = string.Empty;
         public string Color { get; set; } = string.Empty;
         public int CategoryId { get; set; }
-        public long CurrentMileage { get; private set; }
+        public decimal CurrentMileage { get; private set; }
         public DateTime CreateAt { get; private set; } = DateTime.UtcNow;
         public Category Category { get; set; } = default!;
         public bool IsActive { get; private set; } = true;
-        public void UpdateCurrentMileage(long newMileage)
+        public void UpdateCurrentMileage(decimal newMileage)
         {
             if (newMileage < CurrentMileage)
             {
